Server: SMB2: SMB2Session: Fixed KeyNotFoundException when calling the DisconnectTree method

+++ SMBLibrary/Server/ConnectionState/SMB2Session.cs
@@ -91,7 +91,7 @@ namespace SMBLibrary.Server
                 lock (m_openFiles)
                 {
                     List<ulong> fileIDList = new List<ulong>(m_openFiles.Keys);
-                    foreach (ushort fileID in fileIDList)
+                    foreach (ulong fileID in fileIDList)
                     {
                         OpenFileObject openFile = m_openFiles[fileID];
                         if (openFile.TreeID == treeID)
